requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Flask-MySQLdb==2.0.0
python-dotenv==1.0.0
mysqlclient==2.2.7
Flask-Login==0.6.3
Flask-Mail==0.10.0
waitress==3.0.2
gunicorn==21.2.0  # Solo para Linux/Mac
mysql-connector-python==9.2.0
openpyxl==3.1.5
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
bleach==6.1.0